Fix bug where snake can go off screen.
The snake could do off to the top or left.
JPanel.getPreferredSize().getWidth() returns a double, so the snake's
position ends up being a double. The code checks for snake positions
being equal to 0 for the left and top edge collisions, but in clojure
0.0 != 0, so the check never succeeds.

The fix chosen here is to cast the JPanel size to int, so that all
derived values throughout the program are ints.